Added temperature reporting

+++ src/BambikiBackend.Api/Integrations/Firms/Models/AreaCords.cs
@@ -1,3 +1,3 @@
 namespace BambikiBackend.Api.Integrations.Firms.Models;
 
-public record AreaCords(long South, long East, long West, long North);
+public record AreaCords(double South, double East, double West, double North);
